pyproject.toml
[tool.poetry]
name = "zania-assignment"
version = "0.1.0"
description = ""
authors = ["Gourab Sinha <gourab19964u@gmail.com>"]
license = "MIT"
packages = [
    { include = "app" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.5"
uvicorn = {version = "^0.22.0", extras = ["standard"]}
python-multipart = "^0.0.9"
langchain = "^0.1.9"
langchain-openai = "^0.0.7"
langchain-community = "^0.0.24"
faiss-cpu = "^1.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
pypdf = "^5.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload"
clean = "clean_cache:clean_pycache"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true